Extract file upload component from registration form
